requirements.txt
# Date parsing - converts French natural language dates to datetime objects
dateparser>=1.2,<2.0

# Flask API
flask>=3.0,<4.0
flask-sqlalchemy>=3.1,<4.0

# MySQL driver
pymysql>=1.1,<2.0
cryptography>=42.0,<45.0

# LiveKit agent - v1.x API
# [silero] bundles the Silero VAD model
# inference.STT/LLM/TTS are built into livekit-agents, no extra plugin needed
livekit-agents[silero]~=1.4

# HTTP client (agent → API)
httpx>=0.27,<0.29

# Env vars
python-dotenv>=1.0,<2.0
